add correlation_id in pre-filters log
