+++ pkg/web/filter.go
@@ -140,8 +140,9 @@ func (fs Filters) Chain(h Handler) Handler {
 		i := i
 		wrappedFilters[i] = HandlerFunc(func(request *Request) (*Response, error) {
 			params := map[string]interface{}{
-				"path":   request.URL.Path,
-				"method": request.Method,
+				"path":                 request.URL.Path,
+				"method":               request.Method,
+				log.FieldCorrelationID: CorrelationIDFromHeaders(request.Header),
 			}
 
 			log.R(request, fname).WithFields(params).Debug("Entering Filter: ", fs[i].Name())
